Global commands work!
This is kind of a departure from the original branch intent of
whitelists. It turns out that whitelists weren't exactly the
problem--the problem was that things like Google Docs are typically
activating in an about:blank URL rather than simply blacklisting keys.
This adds some global commands to hook up to instead.

They can be edited at: chrome://extensions/shortcuts